feat: only use the fallback font if there is no regular font.
only the symbol and the emoji-font are set as 'fallback' in the backend.

the 'fallback' font is more of a regular font of last resort, and thus set as regular font.

this affects rendering in the backend. the fallback fonts are considered supplemental and not used to determine the font-box size or such things.